#21 - Change notes and AARs to use markdown, allow editing notes (#120)
* Change notes and AARs to use markdown

* Allow editing notes

+++ resources/views/missions/notes/index.blade.php
@@ -1,30 +1,43 @@
 <script>
     $(document).ready(function(e) {
-        var runConvert = function() {
-            $('.mission-comment-item-text').each(function(idx, span) {
-                var jspan = $(span);
-
-                if (!jspan.hasClass('md-converted')) {
-                    span.innerHTML = window.mdconvert.makeHtml(span.innerText);
-                    jspan.addClass('md-converted');
-                }
-            });
-        };
-
-        runConvert();
-
-        $(document).on('click', '.mission-note-delete', function(event) {
+        $(document).on('click', '.mission-note-edit', function(event) {
             var caller = $(this);
+            var id = caller.data('id');
 
             $.ajax({
-                type: 'DELETE',
-                url: '{{ url('/hub/missions/'.$mission->id.'/notes') }}/' + caller.data('id'),
+                type: 'GET',
+                url: '{{ url("/hub/missions/{$mission->id}/notes") }}/' + caller.data('id') + '/edit',
+
                 success: function(data) {
-                    caller.parents('.mission-comment-item').remove();
+                    data = JSON.parse(data);
+                    $('#submit-mission-comment input[name="id"]').val(id);
+                    $('#submit-mission-comment textarea[name="text"]').val(data.text);
+                    $('#submit-mission-comment .mission-aar-textarea').html(data.text);
+                    $('#submit-mission-comment button[type="submit"]').html('Save Changes');
+                    $('#submit-mission-comment textarea[name="text"]').focus();
+                    $('.mission-container').scrollTop(10000);
                 }
             });
 
             event.preventDefault();
+        });
+
+        $(document).on('click', '.mission-note-delete', function(event) {
+            event.preventDefault();
+            var canDelete = confirm("Are you sure you want to delete this?");
+
+            if (canDelete) {
+                var caller = $(this);
+
+                $.ajax({
+                    type: 'DELETE',
+                    url: '{{ url("/hub/missions/{$mission->id}/notes") }}/' + caller.data('id'),
+
+                    success: function(data) {
+                        caller.parents('.mission-comment-item').remove();
+                    }
+                });
+            }
         });
 
         $('#submit-mission-comment').submit(function(event) {
@@ -32,18 +45,19 @@
 
             $.ajax({
                 type: 'POST',
-                url: '{{ url('/hub/missions/'.$mission->id.'/notes') }}',
+                url: '{{ url("/hub/missions/{$mission->id}/notes") }}',
                 data: form.serialize(),
+
                 beforeSend: function() {
                     $('*').blur();
                 },
                 success: function(data) {
-                    $('.mission-comments').append(data);
-                    $('.has-mentions').html('');
+                    $('#submit-mission-comment input[name="id"]').val(-1);
+                    $('#submit-mission-comment textarea[name="text"]').val('');
+                    $('#submit-mission-comment .mission-aar-textarea').html('');
+                    $('#submit-mission-comment button[type="submit"]').html('Submit');
 
-                    runConvert();
-
-                    form[0].reset();
+                    location.reload();
                 }
             });
 
@@ -86,14 +100,12 @@
 <div class="mission-comments-form pull-left w-100">
     <form method="post" id="submit-mission-comment">
         <input type="hidden" name="mentions" id="mentions" value="">
-        <textarea class="form-control" id="submit-mission-comment-text" name="text" style="display:none"></textarea>
+        <input type="hidden" name="id" value="-1'">
 
-        <span
+        <textarea
+            class="form-control-editable form-control mission-aar-textarea m-b-3 m-t-3"
             id="submit-mission-comment-text-real"
-            class="form-control-editable has-mentions mission-aar-textarea form-control m-b-3 m-t-3"
-            contenteditable="plaintext-only"
-            placeholder="Write a note..."
-            for="#submit-mission-comment-text"></span>
+            name="text">{!! (!is_null($mission->draft())) ? $mission->draft()->text : '' !!}</textarea>
 
         <a
             class="pull-left m-l-3"
